Refactor format code

+++ client/src/runtime/eventDelegates.h
@@ -6,7 +6,7 @@ typedef void (* TickDelegate_t)();
 typedef void (* ClientEventDelegate_t)(const char* name, alt::MValueConst** args, uint64_t size);
 typedef void (* ServerEventDelegate_t)(const char* name, alt::MValueConst** args, uint64_t size);
 typedef void (* WebViewEventDelegate_t)(alt::IWebView*, const char* name, alt::MValueConst** args, uint64_t size);
-typedef void (* ConsoleCommandDelegate_t)(const char* name, const char* args[], uint32_t size);
+typedef void (* ConsoleCommandDelegate_t)(const char* name, const char* args[], uint64_t size);
 typedef void (* WebSocketEventDelegate_t)(alt::IWebSocketClient*, const char* name, alt::MValueConst** args, uint64_t size);
 typedef void (* RmlEventDelegate_t)(alt::IRmlElement*, const char* name, alt::IMValueDict* args, uint64_t size);
